Move AsyncTransform code into AsyncPhase.scala
